WIP: new more flexible API
closes #62
