--- examples/.babelrc
@@ -1,18 +1,0 @@
-{
-  "presets": [
-    [
-      "env",
-      {
-        "targets": {
-          "node": "4.5"
-        }
-      }
-    ],
-    "react"
-  ],
-  "plugins": [
-    "babel-macros",
-    "transform-class-properties",
-    "transform-object-rest-spread"
-  ]
-}
