--- examples/README.md
@@ -1,11 +1,0 @@
-# react-autocompletely examples
-
-This is a Next.js app for the examples. This library is a library of primitives
-and that's hard to communicate the value of with regular docs. We need examples
-and lots of them. Each folder in the `pages` directory is a page for examples
-of `react-autocompletely`. We could really use help adding more and more of
-these examples!
-
-**Note**: Unfortunately there were some problems trying to import directly
-from `../src`, so as a pre-build/start script we copy the `../src` to
-`./other/react-autocompletely`.
