--- examples/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "examples",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "cleanup": "rimraf other/react-autocompletely node_modules/.cache && cpy \"./**/*.js\" ../examples/other/react-autocompletely --cwd=\"../src\" --parents",
-    "prestart": "npm run cleanup -s",
-    "start": "next",
-    "prebuild": "npm run cleanup -s",
-    "build": "next build"
-  },
-  "keywords": [],
-  "author": "Kent C. Dodds <kent@doddsfamily.us> (http://kentcdodds.com/)",
-  "license": "MIT",
-  "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-macros": "^0.5.2",
-    "babel-plugin-module-alias": "^1.6.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "cpy-cli": "^1.0.1",
-    "next": "^2.4.7",
-    "preval.macro": "^1.0.1",
-    "rimraf": "^2.6.1"
-  },
-  "dependencies": {
-    "axios": "^0.16.2",
-    "glamor": "^2.20.37",
-    "glamorous": "^3.24.0",
-    "match-sorter": "^1.8.0",
-    "react": "^15.6.1",
-    "react-popper": "^0.7.2",
-    "react-apollo": "^1.4.8",
-    "react-dom": "^15.6.1",
-    "react-instantsearch": "^4.0.9",
-    "throttle-debounce": "^1.0.1"
-  }
-}
