--- examples/pages/react-autocomplete/index.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default Examples
-
-function Examples() {
-  return <div>react-autocomplete examples</div>
-}
