--- examples/pages/react-select/index.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default Examples
-
-function Examples() {
-  return <div>react-select examples</div>
-}
