--- src/constants.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line
-export const AUTOCOMPLETE_CONTEXT = '__autocomplete__'
