+++ stories/.babelrc
@@ -1,0 +1,14 @@
+{
+  "presets": [
+    [
+      "env",
+      {
+        "targets": {
+          "browsers": "ie >= 10"
+        }
+      }
+    ],
+    "react"
+  ],
+  "plugins": ["transform-class-properties", "transform-object-rest-spread"]
+}
